+++ CHANGELOG.md
@@ -40,6 +40,10 @@ The category for changes related to documentation, testing and tooling. Also, fo
 
     *Manuel Puyol*
 
+* Add support for `width: :full` and `height: :full` to System Arguments.
+
+    *Joel Hawksley*
+
 ### Bug fixes
 
 * Update linters to not autocorrect attributes with ERB blocks.